feat: add tests on task-1/ Add mock local storage
