--- components/src/components/input/index.ts
@@ -1,2 +1,0 @@
-import SearchBar from './search';
-export default SearchBar;
