+++ components/src/components/search/index.ts
@@ -1,0 +1,2 @@
+import SearchBar from './search';
+export default SearchBar;
