+++ components/src/storage.ts
@@ -1,0 +1,3 @@
+export function saveToStorage(value: string) {
+  window.localStorage.setItem('the-key', value);
+}
